Update check_hadoop.py
I added a new get_hadoop_version function to get the version of Hadoop using the hadoop version command. This function returns a string representing the version of Hadoop. I then modified the check_hadoop function to call get_hadoop_version and include the Hadoop version in the status message. The script now prints the Hadoop version in the output message along with the status.