Update check_hadoop.py
In this version, I created a HadoopChecker class that contains the get_hadoop_version and check_hadoop methods. The get_hadoop_version method is responsible for getting the Hadoop version and returning it as a string. The check_hadoop method is responsible for checking the status of the Hadoop cluster and returning a tuple of the status and a message.

I also made sure to only call decode('utf-8') once and removed the unnecessary string concatenation in the if statements.

Lastly, I created an instance of the HadoopChecker class and called the check_hadoop method to get the status and message. The script then prints the status and message and exits with the corresponding status code.